pyproject.toml
[project]
name = "polparse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "curl-cffi>=0.10.0",
    "flask>=3.1.0",
    "python-dotenv>=1.1.0",
]
